refactor: 클래스 VisitDto의 이름을 VisitEntity로 변경하여 JPA 엔티티임을 더 명확하게 반영했습니다. 이번 변경을 통해 코드베이스의 의미 전달력이 향상되었습니다. 영향을 받은 파일: 	•	src/main/java/com/realyoungk/sdi/entity/VisitEntity.java (VisitDto.java에서 이름 변경) 	•	src/main/java/com/realyoungk/sdi/service/VisitService.java 	•	src/main/java/com/realyoungk/sdi/repository/VisitJpaRepository.java 	•	src/main/java/com/realyoungk/sdi/repository/VisitRepository.java 기존 VisitDto로 되어 있던 모든 참조와 타입 선언도 VisitEntity로 변경되었습니다.
