+++ src/main/java/com/realyoungk/sdi/repository/VisitRepository.java
@@ -1,7 +1,7 @@
 package com.realyoungk.sdi.repository;
 
 
-import com.realyoungk.sdi.dto.VisitDto;
+import com.realyoungk.sdi.entity.VisitEntity;
 import org.springframework.stereotype.Repository;
 
 import java.util.Date;
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Repository
 public interface VisitRepository {
-    List<VisitDto> findByStartedAtAfterOrderByStartedAtDesc(Date startedAt);
+    List<VisitEntity> findByStartedAtAfterOrderByStartedAtDesc(Date startedAt);
 
-    VisitDto save(VisitDto visitDto);
+    VisitEntity save(VisitEntity visitEntity);
 }
